+++ src/lib.rs
@@ -32,54 +32,54 @@ use std::borrow::Cow;
 pub struct Paillier {}
 
 #[cfg(feature="useramp")]
-pub use arithimpl::rampimpl::BigInteger as BigInteger;
+pub use arithimpl::rampimpl::BigInt as BigInt;
 
 #[cfg(feature="useframp")]
-pub use arithimpl::frampimpl::BigInteger as BigInteger;
+pub use arithimpl::frampimpl::BigInt as BigInt;
 
 #[cfg(feature="usegmp")]
-pub use arithimpl::gmpimpl::BigInteger as BigInteger;
+pub use arithimpl::gmpimpl::BigInt as BigInt;
 
 #[cfg(feature="usenum")]
-pub use arithimpl::numimpl::BigInteger as BigInteger;
+pub use arithimpl::numimpl::BigInt as BigInt;
 
 /// Representation of a keypair from which encryption and decryption keys can be derived.
 pub struct Keypair {
-    pub p: BigInteger,
-    pub q: BigInteger,
+    pub p: BigInt,
+    pub q: BigInt,
 }
 
 /// Public encryption key.
 #[derive(Debug,Clone)]
 pub struct EncryptionKey {
-    n: BigInteger,  // the modulus
-    nn: BigInteger, // the modulus squared
+    n: BigInt,  // the modulus
+    nn: BigInt, // the modulus squared
 }
 
 /// Private decryption key.
 #[derive(Debug,Clone)]
 pub struct DecryptionKey {
-    p: BigInteger,  // first prime
-    q: BigInteger,  // second prime
-    n: BigInteger,  // the modulus (also in public key)
-    nn: BigInteger,
-    pp: BigInteger,
-    pminusone: BigInteger,
-    qq: BigInteger,
-    qminusone: BigInteger,
-    phi: BigInteger,
-    dp: BigInteger,
-    dq: BigInteger,
-    pinv: BigInteger,
-    ppinv: BigInteger,
-    hp: BigInteger,
-    hq: BigInteger,
+    p: BigInt,  // first prime
+    q: BigInt,  // second prime
+    n: BigInt,  // the modulus (also in public key)
+    nn: BigInt,
+    pp: BigInt,
+    pminusone: BigInt,
+    qq: BigInt,
+    qminusone: BigInt,
+    phi: BigInt,
+    dp: BigInt,
+    dq: BigInt,
+    pinv: BigInt,
+    ppinv: BigInt,
+    hp: BigInt,
+    hq: BigInt,
 }
 
 /// Representation of unencrypted message.
 #[derive(Clone,Debug,PartialEq)]
-pub struct RawPlaintext<'b>(Cow<'b, BigInteger>);
+pub struct RawPlaintext<'b>(Cow<'b, BigInt>);
 
 /// Representation of encrypted message.
 #[derive(Clone,Debug,PartialEq)]
-pub struct RawCiphertext<'b>(Cow<'b, BigInteger>);
+pub struct RawCiphertext<'b>(Cow<'b, BigInt>);
